@@ -23,41 +23,41 @@
   },
   "private": true,
   "dependencies": {
-    "@angular/animations": "^12.2.13",
-    "@angular/cdk": "^12.2.13",
-    "@angular/common": "^12.2.13",
-    "@angular/compiler": "^12.2.13",
-    "@angular/core": "^12.2.13",
-    "@angular/forms": "^12.2.13",
-    "@angular/localize": "~12.2.13",
-    "@angular/material": "^12.2.13",
-    "@angular/platform-browser": "^12.2.13",
-    "@angular/platform-browser-dynamic": "^12.2.13",
-    "@angular/platform-server": "^12.2.13",
-    "@angular/router": "^12.2.13",
-    "@elastic/apm-rum-angular": "^2.1.1",
-    "@fortawesome/angular-fontawesome": "^0.8.2",
-    "@fortawesome/fontawesome-svg-core": "^1.2.35",
+    "@angular/animations": "^13.2.2",
+    "@angular/cdk": "^13.2.2",
+    "@angular/common": "^13.2.2",
+    "@angular/compiler": "^13.2.2",
+    "@angular/core": "^13.2.2",
+    "@angular/forms": "^13.2.2",
+    "@angular/localize": "^13.2.2",
+    "@angular/material": "^13.2.2",
+    "@angular/platform-browser": "^13.2.2",
+    "@angular/platform-browser-dynamic": "^13.2.2",
+    "@angular/platform-server": "^13.2.2",
+    "@angular/router": "^13.2.2",
+    "@elastic/apm-rum-angular": "^2.1.4",
+    "@fortawesome/angular-fontawesome": "^0.10.1",
+    "@fortawesome/fontawesome-svg-core": "^1.3.0",
     "@fortawesome/free-brands-svg-icons": "^5.15.3",
     "@fortawesome/free-regular-svg-icons": "^5.15.3",
     "@fortawesome/free-solid-svg-icons": "^5.15.3",
     "@gorniv/ngx-transfer-http": "^2.2.9",
     "@nguniversal/common": "^12.1.3",
     "@nguniversal/express-engine": "^12.1.3",
-    "angular-auth-oidc-client": "^11.6.4",
+    "angular-auth-oidc-client": "^13.1.0",
     "array-flat-polyfill": "^1.0.1",
-    "body-parser": "^1.19.0",
+    "body-parser": "^1.19.2",
     "bootstrap": "^4.6.0",
     "bowser": "2.11.0",
     "cacache": "^15.0.5",
     "compression": "^1.7.4",
-    "core-js": "^3.9.1",
+    "core-js": "^3.21.0",
     "d3-scale-chromatic": "^2.0.0",
     "d3-zoom": "^2.0.0",
     "d3v4": "^4.2.2",
-    "express": "^4.17.1",
+    "express": "^4.17.2",
     "express-csp-header": "^4.0.0",
-    "express-rate-limit": "^6.2.0",
+    "express-rate-limit": "^6.2.1",
     "feature-policy": "^0.6.0",
     "font-awesome": "^4.7.0",
     "helmet": "^4.4.1",
@@ -69,7 +69,7 @@
     "mississippi": "^4.0.0",
     "ngx-bootstrap": "^6.2.0",
     "ngx-countup": "^7.3.3",
-    "nodemailer": "^6.5.0",
+    "nodemailer": "^6.7.2",
     "popper.js": "^1.16.0",
     "referrer-policy": "^1.2.0",
     "reflect-metadata": "^0.1.13",
@@ -81,22 +81,22 @@
     "zone.js": "~0.11.4"
   },
   "devDependencies": {
-    "@angular-devkit/build-angular": "~12.2.13",
-    "@angular/cli": "^12.2.13",
-    "@angular/compiler-cli": "^12.2.13",
-    "@angular/language-service": "^12.2.13",
-    "@babel/cli": "^7.13.10",
-    "@babel/core": "^7.13.10",
-    "@babel/preset-env": "^7.13.10",
+    "@angular-devkit/build-angular": "^13.2.3",
+    "@angular/cli": "^13.2.3",
+    "@angular/compiler-cli": "^13.2.2",
+    "@angular/language-service": "^13.2.2",
+    "@babel/cli": "^7.17.3",
+    "@babel/core": "^7.17.4",
+    "@babel/preset-env": "^7.16.11",
     "@fortawesome/fontawesome-free": "^5.15.3",
     "@ngx-i18nsupport/ngx-i18nsupport": "^1.1.6",
     "@types/d3": "^6.3.0",
     "@types/express": "^4.17.11",
     "@types/jasmine": "~3.6.6",
     "@types/jasminewd2": "^2.0.8",
-    "@types/lodash": "^4.14.168",
-    "@types/lodash-es": "^4.17.5",
-    "@types/node": "^14.14.35",
+    "@types/lodash": "^4.14.178",
+    "@types/lodash-es": "^4.17.6",
+    "@types/node": "^14.18.12",
     "@typescript-eslint/eslint-plugin": "^4.29.3",
     "@typescript-eslint/parser": "^4.29.3",
     "codelyzer": "^6.0.1",
@@ -113,7 +113,7 @@
     "protractor": "~7.0.0",
     "ts-loader": "^8.0.18",
     "ts-node": "~9.1.1",
-    "typescript": "^4.3.5",
-    "webpack-cli": "^4.5.0"
+    "typescript": "^4.5.5",
+    "webpack-cli": "^4.9.2"
   }
 }
