@@ -6,41 +6,36 @@
 //  :license: MIT
 
 import { HttpClient } from '@angular/common/http';
-import { APP_INITIALIZER, NgModule } from '@angular/core';
+import { NgModule } from '@angular/core';
 import {
   AuthModule,
-  OidcConfigService,
-  OidcSecurityService,
+  StsConfigHttpLoader,
+  StsConfigLoader,
 } from 'angular-auth-oidc-client';
-import { map, switchMap } from 'rxjs/operators';
+import { map } from 'rxjs/operators';
 
-export function configureAuth(
-  oidcConfigService: OidcConfigService,
-  httpClient: HttpClient
-) {
-  const setupAction$ = httpClient
-    .get<any>(`assets/config/auth_config.json`)
-    .pipe(
-      map((authConfig) => {
-        return authConfig;
-      }),
-      switchMap((config) => oidcConfigService.withConfig(config))
-    );
+export function httpLoaderFactory(httpClient: HttpClient) {
+  const config$ = httpClient.get<any>(`assets/config/auth_config.json`).pipe(
+    map((customConfig) => {
+      return {
+        authority: customConfig.stsServer,
+        ...customConfig,
+      };
+    })
+  );
 
-  return () => setupAction$.toPromise();
+  return new StsConfigHttpLoader(config$);
 }
 
 @NgModule({
-  imports: [AuthModule.forRoot()],
-  providers: [
-    OidcSecurityService,
-    OidcConfigService,
-    {
-      provide: APP_INITIALIZER,
-      useFactory: configureAuth,
-      deps: [OidcConfigService, HttpClient],
-      multi: true,
-    },
+  imports: [
+    AuthModule.forRoot({
+      loader: {
+        provide: StsConfigLoader,
+        useFactory: httpLoaderFactory,
+        deps: [HttpClient],
+      },
+    }),
   ],
   exports: [AuthModule],
 })
