@@ -32,35 +32,34 @@ import { WINDOW_PROVIDERS } from '../shared/services/window.service';
 import { TooltipModule } from 'ngx-bootstrap/tooltip';
 
 @NgModule({
-  declarations: [
-    LayoutComponent,
-    HeaderComponent,
-    FooterComponent,
-    ErrorModalComponent,
-    ReviewComponent,
-    BetaInfoComponent,
-  ],
-  imports: [
-    CommonModule,
-    SharedModule,
-    AppRoutingModule,
-    MatSelectModule,
-    MatMenuModule,
-    MatIconModule,
-    MatButtonModule,
-    MatDialogModule,
-    MatRadioModule,
-    MatFormFieldModule,
-    MatInputModule,
-    MatCheckboxModule,
-    MatProgressSpinnerModule,
-    FormsModule,
-    FontAwesomeModule,
-    // ClickOutsideModule,
-    TooltipModule.forRoot(),
-  ],
-  exports: [LayoutComponent],
-  entryComponents: [ReviewComponent],
-  providers: [WINDOW_PROVIDERS],
+    declarations: [
+        LayoutComponent,
+        HeaderComponent,
+        FooterComponent,
+        ErrorModalComponent,
+        ReviewComponent,
+        BetaInfoComponent,
+    ],
+    imports: [
+        CommonModule,
+        SharedModule,
+        AppRoutingModule,
+        MatSelectModule,
+        MatMenuModule,
+        MatIconModule,
+        MatButtonModule,
+        MatDialogModule,
+        MatRadioModule,
+        MatFormFieldModule,
+        MatInputModule,
+        MatCheckboxModule,
+        MatProgressSpinnerModule,
+        FormsModule,
+        FontAwesomeModule,
+        // ClickOutsideModule,
+        TooltipModule.forRoot(),
+    ],
+    exports: [LayoutComponent],
+    providers: [WINDOW_PROVIDERS]
 })
 export class LayoutModule {}
