@@ -166,157 +166,149 @@ import { MatSortModule } from '@angular/material/sort';
 import { HandleInfrastructureLinkPipe } from './pipes/handle-infrastructure-link.pipe';
 
 @NgModule({
-  declarations: [
-    HomePageComponent,
-    SearchBarComponent,
-    ResultsComponent,
-    SinglePublicationComponent,
-    PublicationsComponent,
-    PersonsComponent,
-    ResultTabComponent,
-    SortComponent,
-    PaginationComponent,
-    BreadcrumbComponent,
-    FundingsComponent,
-    SingleFundingComponent,
-    ActiveFiltersComponent,
-    OrganizationsComponent,
-    SingleOrganizationComponent,
-    VisualisationComponent,
-    RelatedResultsComponent,
-    SearchResultsComponent,
-    EmptyResultComponent,
-    ListItemComponent,
-    SunburstComponent,
-    TreemapComponent,
-    HighlightSearch,
-    TooltipComponent,
-    SuggestComponent,
-    NewsComponent,
-    LinksPipe,
-    InfrastructuresComponent,
-    ShareComponent,
-    SingleInfrastructureComponent,
-    OrcidComponent,
-    FiltersComponent,
-    CounterPipe,
-    FilterItemPipe,
-    FilterSumPipe,
-    SafeUrlPipe,
-    ResearchInnovationSystemComponent,
-    FiguresComponent,
-    ScrollSpyDirective,
-    CleanCitationPipe,
-    ReplaceSpacePipe,
-    SingleFigureComponent,
-    RelatedLinksComponent,
-    FilterListComponent,
-    ServiceInfoComponent,
-    PrivacyComponent,
-    AccessibilityComponent,
-    NewsCardComponent,
-    SitemapComponent,
-    TabItemComponent,
-    NewsPaginationComponent,
-    CarouselComponent,
-    ResultCountComponent,
-    BarComponent,
-    FigureFiltersComponent,
-    FiguresInfoComponent,
-    DatasetsComponent,
-    SingleDatasetComponent,
-    ExternalLinksComponent,
-    BannerComponent,
-    LatestNewsComponent,
-    NewsResultsComponent,
-    PieComponent,
-    ConvertToArrayPipe,
-    FundingCallsComponent,
-    SingleFundingCallComponent,
-    FundingCallResultsComponent,
-    FundingCallPreviewComponent,
-    HandleLinkDisplayPipe,
-    IsUrlPipe,
-    OrganizationVisualisationsComponent,
-    OrganizationInformationComponent,
-    OrganizationSubUnitsComponent,
-    FilterEmptyFieldPipe,
-    HandleLinkDisplayPipe,
-    IsUrlPipe,
-    OrganizationVisualisationsComponent,
-    OrganizationInformationComponent,
-    OrganizationSubUnitsComponent,
-    FilterEmptyFieldPipe,
-    PublicationLinksComponent,
-    DatasetAuthorComponent,
-    TkiReportsComponent,
-    HandleInfrastructureLinkPipe,
-  ],
-  imports: [
-    PortalRoutingModule,
-    CommonModule,
-    TypeaheadModule.forRoot(),
-    FormsModule,
-    ReactiveFormsModule,
-    MatIconModule,
-    MatInputModule,
-    MatAutocompleteModule,
-    MatChipsModule,
-    MatFormFieldModule,
-    MatCheckboxModule,
-    MatCardModule,
-    MatDatepickerModule,
-    MatNativeDateModule,
-    MatRadioModule,
-    MatSelectModule,
-    MatListModule,
-    MatTabsModule,
-    MatProgressSpinnerModule,
-    MatExpansionModule,
-    MatPaginatorModule,
-    MatButtonModule,
-    MatSlideToggleModule,
-    MatDialogModule,
-    MatSnackBarModule,
-    MatMenuModule,
-    ScrollingModule,
-    ClipboardModule,
-    CountUpModule,
-    FontAwesomeModule,
-    TransferHttpCacheModule,
-    ModalModule.forRoot(),
-    SharedModule,
-    A11yModule,
-    TooltipModule.forRoot(),
-    MatTableModule,
-    MatSortModule,
-  ],
-  exports: [DatasetAuthorComponent],
-  providers: [
-    SearchService,
-    Title,
-    Meta,
-    AutosuggestService,
-    WINDOW_PROVIDERS,
-    AppConfigService,
-    Location,
-    {
-      provide: LocationStrategy,
-      useClass: PathLocationStrategy,
-    },
-    {
-      provide: MAT_SNACK_BAR_DEFAULT_OPTIONS,
-      useValue: { duration: 3000 },
-    },
-  ],
-  entryComponents: [
-    PublicationsComponent,
-    PersonsComponent,
-    FundingsComponent,
-    InfrastructuresComponent,
-    OrganizationsComponent,
-    EmptyResultComponent,
-  ],
+    declarations: [
+        HomePageComponent,
+        SearchBarComponent,
+        ResultsComponent,
+        SinglePublicationComponent,
+        PublicationsComponent,
+        PersonsComponent,
+        ResultTabComponent,
+        SortComponent,
+        PaginationComponent,
+        BreadcrumbComponent,
+        FundingsComponent,
+        SingleFundingComponent,
+        ActiveFiltersComponent,
+        OrganizationsComponent,
+        SingleOrganizationComponent,
+        VisualisationComponent,
+        RelatedResultsComponent,
+        SearchResultsComponent,
+        EmptyResultComponent,
+        ListItemComponent,
+        SunburstComponent,
+        TreemapComponent,
+        HighlightSearch,
+        TooltipComponent,
+        SuggestComponent,
+        NewsComponent,
+        LinksPipe,
+        InfrastructuresComponent,
+        ShareComponent,
+        SingleInfrastructureComponent,
+        OrcidComponent,
+        FiltersComponent,
+        CounterPipe,
+        FilterItemPipe,
+        FilterSumPipe,
+        SafeUrlPipe,
+        ResearchInnovationSystemComponent,
+        FiguresComponent,
+        ScrollSpyDirective,
+        CleanCitationPipe,
+        ReplaceSpacePipe,
+        SingleFigureComponent,
+        RelatedLinksComponent,
+        FilterListComponent,
+        ServiceInfoComponent,
+        PrivacyComponent,
+        AccessibilityComponent,
+        NewsCardComponent,
+        SitemapComponent,
+        TabItemComponent,
+        NewsPaginationComponent,
+        CarouselComponent,
+        ResultCountComponent,
+        BarComponent,
+        FigureFiltersComponent,
+        FiguresInfoComponent,
+        DatasetsComponent,
+        SingleDatasetComponent,
+        ExternalLinksComponent,
+        BannerComponent,
+        LatestNewsComponent,
+        NewsResultsComponent,
+        PieComponent,
+        ConvertToArrayPipe,
+        FundingCallsComponent,
+        SingleFundingCallComponent,
+        FundingCallResultsComponent,
+        FundingCallPreviewComponent,
+        HandleLinkDisplayPipe,
+        IsUrlPipe,
+        OrganizationVisualisationsComponent,
+        OrganizationInformationComponent,
+        OrganizationSubUnitsComponent,
+        FilterEmptyFieldPipe,
+        HandleLinkDisplayPipe,
+        IsUrlPipe,
+        OrganizationVisualisationsComponent,
+        OrganizationInformationComponent,
+        OrganizationSubUnitsComponent,
+        FilterEmptyFieldPipe,
+        PublicationLinksComponent,
+        DatasetAuthorComponent,
+        TkiReportsComponent,
+        HandleInfrastructureLinkPipe,
+    ],
+    imports: [
+        PortalRoutingModule,
+        CommonModule,
+        TypeaheadModule.forRoot(),
+        FormsModule,
+        ReactiveFormsModule,
+        MatIconModule,
+        MatInputModule,
+        MatAutocompleteModule,
+        MatChipsModule,
+        MatFormFieldModule,
+        MatCheckboxModule,
+        MatCardModule,
+        MatDatepickerModule,
+        MatNativeDateModule,
+        MatRadioModule,
+        MatSelectModule,
+        MatListModule,
+        MatTabsModule,
+        MatProgressSpinnerModule,
+        MatExpansionModule,
+        MatPaginatorModule,
+        MatButtonModule,
+        MatSlideToggleModule,
+        MatDialogModule,
+        MatSnackBarModule,
+        MatMenuModule,
+        ScrollingModule,
+        ClipboardModule,
+        CountUpModule,
+        FontAwesomeModule,
+        TransferHttpCacheModule,
+        ModalModule.forRoot(),
+        SharedModule,
+        A11yModule,
+        TooltipModule.forRoot(),
+        MatTableModule,
+        MatSortModule,
+    ],
+    exports: [DatasetAuthorComponent],
+    providers: [
+        SearchService,
+        Title,
+        Meta,
+        AutosuggestService,
+        WINDOW_PROVIDERS,
+        AppConfigService,
+        Location,
+        {
+            provide: LocationStrategy,
+            useClass: PathLocationStrategy,
+        },
+        {
+            provide: MAT_SNACK_BAR_DEFAULT_OPTIONS,
+            useValue: { duration: 3000 },
+        },
+    ]
 })
 export class PortalModule {
   startPage;
@@ -415,6 +407,6 @@ export class PortalModule {
         }
       });
     // Add global icons
-    library.addIcons(faExternalLinkAlt, faInfoCircle);
+    library.addIcons(faExternalLinkAlt as any, faInfoCircle as any);
   }
 }
